fix: Remove useless warning about superfluous collection members
This warning currently prevents users from easily piping dicts into `Collection.validate` that have extra members. This is often a useful shortcut that allows for elegant and practical code. Additionally, it is hard to imagine when a user would be happy about this warning. Even if it triggers, the user is getting what they ordered: a validated collection with the right members. There is no real opportunity for suprising / undesirable results here.